requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
narwhals==1.8.0
numpy==1.21.0
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
preshed==3.0.9
protobuf==5.28.1
pyarrow==17.0.0
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.38.0
tenacity==8.5.0
thinc==8.2.5
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
